ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
numpy==1.26.4
